Move more helper funcs to canvasUtils
