+++ src/App.js
@@ -18,7 +18,16 @@ import usePanAndZoom, {
   positionToTimestamp,
   timestampToPosition,
 } from './usePanAndZoom';
-import { getCanvasContext, getTimeTickInterval, cachedFlamegraphTextWidths, trimFlamegraphText } from './canvasUtils';
+
+import { 
+  getCanvasContext,
+  getTimeTickInterval,
+  cachedFlamegraphTextWidths,
+  trimFlamegraphText,
+  getHoveredEvent,
+  getPriorityHeight,
+} from './canvasUtils';
+
 import prettyMilliseconds from 'pretty-ms';
 import { getBatchRange } from './utils';
 import useInteractiveEvents from './useInteractiveEvents';
@@ -80,139 +89,6 @@ type ContextMenuContextData = {|
   hoveredEvent: ReactHoverContextInfo | null,
   state: PanAndZoomState,
 |};
-
-// TODO: Move to canvasutils
-function getHoveredEvent(
-  schedulerCanvasHeight: number,
-  data: ReactProfilerData | null,
-  flamechart: FlamechartData | null,
-  state: PanAndZoomState
-): ReactHoverContextInfo | null {
-  const { canvasMouseX, canvasMouseY, offsetY } = state;
-
-  if (canvasMouseX < LABEL_FIXED_WIDTH || canvasMouseY < HEADER_HEIGHT_FIXED) {
-    return null;
-  }
-
-  if (canvasMouseY + offsetY < schedulerCanvasHeight) {
-    if (data != null) {
-      let adjustedCanvasMouseY = canvasMouseY - HEADER_HEIGHT_FIXED + offsetY;
-      let priorityMinY = HEADER_HEIGHT_FIXED;
-      let priorityIndex = null;
-      let priority: ReactPriority = 'unscheduled';
-      for (let index = 0; index < REACT_PRIORITIES.length; index++) {
-        priority = REACT_PRIORITIES[index];
-
-        const priorityHeight = getPriorityHeight(data, priority);
-        if (
-          adjustedCanvasMouseY >= priorityMinY &&
-          adjustedCanvasMouseY <= priorityMinY + priorityHeight
-        ) {
-          priorityIndex = index;
-          break;
-        }
-        priorityMinY += priorityHeight;
-      }
-
-      if (priorityIndex === null) {
-        return null;
-      }
-
-      const baseY = priorityMinY - offsetY;
-      const eventMinY = baseY + REACT_GUTTER_SIZE / 2;
-      const eventMaxY = eventMinY + REACT_EVENT_SIZE + REACT_GUTTER_SIZE;
-      const measureMinY = eventMaxY;
-      const measureMaxY = measureMinY + REACT_WORK_SIZE + REACT_GUTTER_SIZE;
-
-      let events = null,
-        measures = null;
-      if (canvasMouseY >= eventMinY && canvasMouseY <= eventMaxY) {
-        events = data[priority].events;
-      } else if (canvasMouseY >= measureMinY && canvasMouseY <= measureMaxY) {
-        measures = data[priority].measures;
-      }
-
-      if (events !== null) {
-        for (let index = events.length - 1; index >= 0; index--) {
-          const event = events[index];
-          const { timestamp } = event;
-
-          const eventX = timestampToPosition(timestamp, state);
-          const startX = eventX - REACT_EVENT_SIZE / 2;
-          const stopX = eventX + REACT_EVENT_SIZE / 2;
-          if (canvasMouseX >= startX && canvasMouseX <= stopX) {
-            return {
-              event,
-              flamechartNode: null,
-              measure: null,
-              priorityIndex,
-              data,
-            };
-          }
-        }
-      } else if (measures !== null) {
-        // Because data ranges may overlap, wew ant to find the last intersecting item.
-        // This will always be the one on "top" (the one the user is hovering over).
-        for (let index = measures.length - 1; index >= 0; index--) {
-          const measure = measures[index];
-          const { duration, timestamp } = measure;
-
-          const pointerTime = positionToTimestamp(canvasMouseX, state);
-
-          if (pointerTime >= timestamp && pointerTime <= timestamp + duration) {
-            return {
-              event: null,
-              flamechartNode: null,
-              measure,
-              priorityIndex,
-              data,
-            };
-          }
-        }
-      }
-    }
-  } else {
-    if (flamechart !== null) {
-      const layerIndex = Math.floor(
-        (canvasMouseY + offsetY - HEADER_HEIGHT_FIXED - schedulerCanvasHeight) /
-          FLAMECHART_FRAME_HEIGHT
-      );
-      const layer = flamechart.layers[layerIndex];
-
-      if (layer != null) {
-        let startIndex = 0;
-        let stopIndex = layer.length - 1;
-        while (startIndex <= stopIndex) {
-          const currentIndex = Math.floor((startIndex + stopIndex) / 2);
-          const flamechartNode = layer[currentIndex];
-
-          const { end, start } = flamechartNode;
-
-          const width = durationToWidth((end - start) / 1000, state);
-          const x = Math.floor(timestampToPosition(start / 1000, state));
-
-          if (x <= canvasMouseX && x + width >= canvasMouseX) {
-            return {
-              event: null,
-              flamechartNode,
-              measure: null,
-              priorityIndex: null,
-              data,
-            };
-          }
-
-          if (x > canvasMouseX) {
-            stopIndex = currentIndex - 1;
-          } else {
-            startIndex = currentIndex + 1;
-          }
-        }
-      }
-    }
-  }
-
-  return null;
-}
 
 const renderReact = ({
   baseY,
@@ -650,46 +526,6 @@ const renderCanvas = memoize(
     }
   }
 );
-
-const cachedPriorityHeights = new Map();
-const getPriorityHeight = (
-  data: ReactProfilerData,
-  priority: ReactPriority
-): number => {
-  if (cachedPriorityHeights.has(priority)) {
-    // We know the value must be present because we've just checked.
-    return ((cachedPriorityHeights.get(priority): any): number);
-  } else {
-    const numMeasures = data[priority].maxNestedMeasures;
-    const events = data[priority].events;
-
-    let priorityHeight = 0;
-    if (numMeasures > 0 && events.length > 0) {
-      priorityHeight =
-        REACT_GUTTER_SIZE +
-        REACT_EVENT_SIZE +
-        REACT_WORK_SIZE * numMeasures +
-        REACT_GUTTER_SIZE * numMeasures +
-        REACT_PRIORITY_BORDER_SIZE;
-    } else if (numMeasures > 0) {
-      priorityHeight =
-        REACT_GUTTER_SIZE +
-        REACT_WORK_SIZE * numMeasures +
-        REACT_GUTTER_SIZE * numMeasures +
-        REACT_PRIORITY_BORDER_SIZE;
-    } else if (events.length > 0) {
-      priorityHeight =
-        REACT_GUTTER_SIZE +
-        REACT_EVENT_SIZE +
-        REACT_GUTTER_SIZE +
-        REACT_PRIORITY_BORDER_SIZE;
-    }
-
-    cachedPriorityHeights.set(priority, priorityHeight);
-
-    return priorityHeight;
-  }
-};
 
 function App() {
   const [data, setData] = useState<ReactProfilerData | null>(null);
